src/requirements.txt
PyYAML>=3.11
Jinja2>=2.8
